python/requirements.txt
streamlit>=1.32.0
supabase>=2.4.0
python-dotenv>=1.0.0
fpdf2>=2.7.9
plotly>=5.20.0
pandas>=2.2.0
